Move model into separate file
